refactor: move types to lib/db & rename schema/users to schema/profiles
