+++ src/lib/components/NavBar/NotificationsMenu.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { NotificationTypes } from '$lib/db/schema/notifications';
-	import type { FullProfile } from '$lib/db/schema/types';
+	import type { FullProfile } from '$lib/db/types';
 	import Button from '../Button.svelte';
 	import { fly } from 'svelte/transition';
 	import FriendRequestNotification from './notifications/FriendRequestNotification.svelte';
