+++ src/lib/db/schema/profiles.ts
@@ -24,7 +24,7 @@ import {
 	defaultMusicWidget
 } from '../../widgets/default-widgets';
 import { authenticatedRole, authUid, authUsers } from 'drizzle-orm/supabase';
-import type { Profile } from './types';
+import type { Profile } from '../types';
 
 export const USERNAME_REGEX = /^([a-zA-Z0-9_]{2,24})$/;
