+++ src/params/profile.ts
@@ -1,4 +1,4 @@
-import { USERNAME_REGEX } from '$lib/db/schema/users';
+import { USERNAME_REGEX } from '$lib/db/schema/profiles';
 
 export function match(param) {
 	return USERNAME_REGEX.test(param);
