+++ src/routes/api/profile/edit-connection.ts
@@ -2,7 +2,7 @@ import { fail, redirect } from '@sveltejs/kit';
 import type { RequestEvent } from './$types';
 import { connectionProviders } from '$lib/widgets/connections';
 import { db } from '$lib/db';
-import { connections } from '$lib/db/schema/users';
+import { connections } from '$lib/db/schema/profiles';
 import { eq } from 'drizzle-orm';
 
 export async function editConnection({
